+++ google-cloud-trace/lib/google/cloud/trace.rb
@@ -22,10 +22,10 @@ require "google/cloud/trace/middleware"
 require "google/cloud/trace/notifications"
 require "google/cloud/trace/project"
 require "google/cloud/trace/result_set"
-require "google/cloud/trace/sampling"
 require "google/cloud/trace/service"
 require "google/cloud/trace/span"
 require "google/cloud/trace/span_kind"
+require "google/cloud/trace/time_sampler"
 require "google/cloud/trace/trace_record"
 require "google/cloud/trace/utils"
 
@@ -68,6 +68,12 @@ module Google
     # to the "Trace" section. It also integrates with Google App Engine
     # Flexible and Google Container Engine to provide additional information
     # for applications hosted in those environments.
+    #
+    # Note that not all requests will have traces. By default, the library will
+    # sample about one trace every ten seconds per Ruby process, to prevent
+    # heavily used applications from reporting too much data. It will also
+    # omit certain requests used by Google App Engine for health checking. See
+    # {Google::Cloud::Trace::TimeSampler} for more details.
     #
     # ### Using instrumentation with Ruby on Rails
     #
